+++ python/src/alayalite/laser/preprocess.py
@@ -1,0 +1,6 @@
+from sklearn import preprocessing
+
+
+def normalize(X):
+    X = preprocessing.normalize(X, axis=1, norm="l2")
+    return X